# internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: armory_api
  description: >
    Бэкенд для armory_web (ПР4 «Подключение REST API»). Реализует тот же
    контракт, что учебный mock-сервер (КОНТРАКТ-API.md), для домена
    оружейного магазина вместо библиотеки.
  version: "1.0"
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /weapons:
    get:
      tags: [weapons]
      summary: Список оружия
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: manufacturerId
          in: query
          schema: { type: integer }
        - name: designerId
          in: query
          schema: { type: integer }
        - name: yearFrom
          in: query
          schema: { type: integer }
        - name: yearTo
          in: query
          schema: { type: integer }
        - name: sort
          in: query
          description: "Поле и направление через запятую, например year,desc. Поля: name, year, price."
          schema: { type: string, example: "name,asc" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Deleted"
      responses:
        "200":
          description: Страница результатов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WeaponPage" }
    post:
      tags: [weapons]
      summary: Создать оружие
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WeaponInput" }
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Weapon" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /manufacturers:
    get:
      tags: [manufacturers]
      summary: Список производителей
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: sort
          in: query
          description: "Поля: name, country, founded."
          schema: { type: string, example: "name,asc" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Deleted"
      responses:
        "200":
          description: Страница результатов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManufacturerPage" }
    post:
      tags: [manufacturers]
      summary: Создать производителя
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ManufacturerInput" }
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Manufacturer" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /categories:
    get:
      tags: [categories]
      summary: Список категорий
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: sort
          in: query
          description: "Поле: name."
          schema: { type: string, example: "name,asc" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Deleted"
      responses:
        "200":
          description: Страница результатов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryPage" }
    post:
      tags: [categories]
      summary: Создать категорию
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /designers:
    get:
      tags: [designers]
      summary: Список конструкторов
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: sort
          in: query
          description: "Поля: fullName, country, activeSince."
          schema: { type: string, example: "fullName,asc" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Deleted"
      responses:
        "200":
          description: Страница результатов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DesignerPage" }
    post:
      tags: [designers]
      summary: Создать конструктора
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DesignerInput" }
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Designer" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /clients:
    get:
      tags: [clients]
      summary: Список покупателей
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: sort
          in: query
          description: "Поля: fullName, email."
          schema: { type: string, example: "fullName,asc" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Deleted"
      responses:
        "200":
          description: Страница результатов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientPage" }
    post:
      tags: [clients]
      summary: Создать покупателя
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientInput" }
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  parameters:
...
    Search:
      name: search
      in: query
      description: Поиск подстрокой без учёта регистра
      schema: { type: string }
    Page:
      name: page
      in: query
      description: Номер страницы, с 1
      schema: { type: integer, default: 1, minimum: 1 }
    Size:
      name: size
      in: query
      description: Размер страницы, по умолчанию 10, максимум 100
      schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
    Deleted:
      name: deleted
      in: query
      description: "'1' — включать логически удалённые записи"
      schema: { type: string, enum: ["1"] }
...
  responses:
...
    ValidationFailed:
      description: Ошибка валидации полей
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
...
  schemas:
...
    ValidationError:
      type: object
      properties:
        message: { type: string, example: "Ошибка валидации" }
        errors:
          type: object
          additionalProperties: { type: string }
          example: { sku: "Артикул «KR-P9-001» уже используется" }
...
    Manufacturer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 120 }
        country: { type: string, maxLength: 60 }
        founded: { type: integer, minimum: 1300, maximum: 2100 }
        deletedAt: { type: string, format: date-time, nullable: true }
    ManufacturerInput:
      type: object
      required: [name, country, founded]
      properties:
        name: { type: string, maxLength: 120 }
        country: { type: string, maxLength: 60 }
        founded: { type: integer, minimum: 1300, maximum: 2100 }
    ManufacturerPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Manufacturer" }
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 60 }
        deletedAt: { type: string, format: date-time, nullable: true }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 60 }
    CategoryPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Category" }
    Designer:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string, maxLength: 120 }
        country: { type: string, maxLength: 60 }
        activeSince: { type: integer, minimum: 1300, maximum: 2100 }
        deletedAt: { type: string, format: date-time, nullable: true }
    DesignerInput:
      type: object
      required: [fullName, country, activeSince]
      properties:
        fullName: { type: string, maxLength: 120 }
        country: { type: string, maxLength: 60 }
        activeSince: { type: integer, minimum: 1300, maximum: 2100 }
    DesignerPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Designer" }
    Client:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string, maxLength: 120 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 5, maxLength: 20 }
        licenseNumber: { type: string, minLength: 3, maxLength: 30 }
        licenseIssuedAt: { type: string, format: date-time }
        licenseExpiresAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
    ClientInput:
      type: object
      required: [fullName, email, phone, licenseNumber, licenseIssuedAt, licenseExpiresAt]
      properties:
        fullName: { type: string, maxLength: 120 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 5, maxLength: 20 }
        licenseNumber: { type: string, minLength: 3, maxLength: 30 }
        licenseIssuedAt: { type: string, format: date-time }
        licenseExpiresAt: { type: string, format: date-time, description: "Строго позже licenseIssuedAt" }
    ClientPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Client" }
    Weapon:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 120 }
        sku: { type: string, minLength: 3, maxLength: 40 }
        year: { type: integer, minimum: 1870, maximum: 2100 }
        caliber: { type: string, maxLength: 30 }
        manufacturerId: { type: integer }
        categoryIds:
          type: array
          items: { type: integer }
        designerIds:
          type: array
          items: { type: integer }
        price: { type: integer, minimum: 1 }
        stockTotal: { type: integer, minimum: 1 }
        stockAvailable: { type: integer, minimum: 0, description: "Не больше stockTotal" }
        deletedAt: { type: string, format: date-time, nullable: true }
    WeaponInput:
      type: object
      required: [name, sku, year, caliber, manufacturerId, categoryIds, designerIds, price, stockTotal, stockAvailable]
      properties:
        name: { type: string, maxLength: 120 }
        sku: { type: string, minLength: 3, maxLength: 40 }
        year: { type: integer, minimum: 1870, maximum: 2100 }
        caliber: { type: string, maxLength: 30 }
        manufacturerId: { type: integer }
        categoryIds:
          type: array
          items: { type: integer }
          minItems: 1
        designerIds:
          type: array
          items: { type: integer }
          minItems: 1
        price: { type: integer, minimum: 1 }
        stockTotal: { type: integer, minimum: 1 }
        stockAvailable: { type: integer, minimum: 0 }
    WeaponPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Weapon" }
    PageEnvelope:
      type: object
      properties:
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }